[refactor] setting bar에서 buttons컴포넌트 분리
